// src/Components/Interfaces/CraftsmenDataInterface.ts
export interface CraftsmenInterface{
    userId: string;
    registrationDate:string;
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: string;
    password: string;
    personalIdNumber: string;
    city: string;
    district: string;
    profession: string;
    experience: string;
    price: string;
    verificationStatus: string;
    minCallFee: string;
    moreInformation: string;
    profilePicture: string;
}